arm64: wait for transmit completion before next character transmission
Previous transmission must be completed before next character to be
transmitted, otherwise TX buffer may saturate and we will not see all
the characters on screen.

We pass Port's LSR address and the value to be checked on that address
for tx port to be empty. This port can be 8/16/32 bit, however we
currently always treat them as 8 bit. Can pass another argument if
someone complains in future.

Signed-off-by: Pratyush Anand <panand@redhat.com>