arm64: crash: Fix mem reservation for more than one crash regions
We have one platform where System RAM is as follows:

00000000-396fffff : System RAM
39dd0000-3edbffff : System RAM
3edf0000-3fbfffff : System RAM
1040000000-1ffbffffff : System RAM
2000000000-2ffbffffff : System RAM
41000000000-41ffbffffff : System RAM
42000000000-42ffbffffff : System RAM

Here,memblock_find_in_range() for crashkernel fails as
ARCH_LOW_ADDRESS_LIMIT turns out to be 0xffffffff.

For such platform we try to find memblock in the range of 0 and
PHYS_MAX, and at the same time allocate some minimal memory low
address range for DMA needs. So, now we can have two regions for "Crash
kernel".

This patch updates code to add more than one usable range. We would not
yet expect above platform to work, as this is just a preparation patch.
However, we should see that this patch does not break any exiting
upstream supported platform.

Signed-off-by: Pratyush Anand <panand@redhat.com>